+++ test2/amount-conversion.test.ts
@@ -1,10 +1,10 @@
 import * as test from "tape";
 import { onlySkip } from "./test-data-utils";
-import * as TestData from "./amount-duration-conversion-data";
+import * as DurationConversion from "./duration-conversion-data";
 import * as Amount from "../src/amount";
 
 test("amount_duration_test", t => {
-  onlySkip(TestData.tests).forEach(item => {
+  onlySkip(DurationConversion.tests).forEach(item => {
     t.test(item.name, st => {
       const amountToTest = Amount.create(item.fromValue, item.fromUnit);
       const actualToValue: number = Amount.valueAs(item.toUnit, amountToTest);
